[#38]Refactor: Change ThemePlaceDto for querydsl
