+++ YAPP-BE/src/main/java/yapp/domain/townMap/dto/InfraPlaceDto.java
